Use pre-assembled plugin archive from trigger build (#47)
* Use pre-assembled plugin archive from trigger build

* Fix path to pre-assembled plugin tar

* Fix path to pre-assembled plugin tar

* No need to check aggregator pom check if using a pre-assembled hpi archive

+++ run.sh
@@ -112,21 +112,24 @@ if [ ! -d "${AGGREGATOR_DIR}" ]; then
     echo ""
     exit 1
 fi
-pushd "${AGGREGATOR_DIR}"
-AGGREGATOR_GROUP_ID=`echo -e 'setns x=http://maven.apache.org/POM/4.0.0\ncat /x:project/x:parent/x:groupId/text()' | xmllint --shell pom.xml | grep -v /`
-AGGREGATOR_ARTIFACT_ID=`echo -e 'setns x=http://maven.apache.org/POM/4.0.0\ncat /x:project/x:artifactId/text()' | xmllint --shell pom.xml | grep -v /`
-popd
-if [ "${AGGREGATOR_GROUP_ID}" != "io.jenkins.blueocean" ] || [ "${AGGREGATOR_ARTIFACT_ID}" != "blueocean" ]; then
-    echo ""
-    echo " *********************************************************************"
-    echo "    The location specified for the aggregator plugin does not appear"
-    echo "    to be correct. Check the supplied parameter and make sure it"
-    echo "    points to the aggregator plugin."
-    echo "    > groupId:    ${AGGREGATOR_GROUP_ID}"
-    echo "    > artifactId: ${AGGREGATOR_ARTIFACT_ID}"
-    echo " *********************************************************************"
-    echo ""
-    exit 1
+
+if [ ! -f "${AGGREGATOR_DIR}/.pre-assembly" ]; then
+    pushd "${AGGREGATOR_DIR}"
+    AGGREGATOR_GROUP_ID=`echo -e 'setns x=http://maven.apache.org/POM/4.0.0\ncat /x:project/x:parent/x:groupId/text()' | xmllint --shell pom.xml | grep -v /`
+    AGGREGATOR_ARTIFACT_ID=`echo -e 'setns x=http://maven.apache.org/POM/4.0.0\ncat /x:project/x:artifactId/text()' | xmllint --shell pom.xml | grep -v /`
+    popd
+    if [ "${AGGREGATOR_GROUP_ID}" != "io.jenkins.blueocean" ] || [ "${AGGREGATOR_ARTIFACT_ID}" != "blueocean" ]; then
+        echo ""
+        echo " *********************************************************************"
+        echo "    The location specified for the aggregator plugin does not appear"
+        echo "    to be correct. Check the supplied parameter and make sure it"
+        echo "    points to the aggregator plugin."
+        echo "    > groupId:    ${AGGREGATOR_GROUP_ID}"
+        echo "    > artifactId: ${AGGREGATOR_ARTIFACT_ID}"
+        echo " *********************************************************************"
+        echo ""
+        exit 1
+    fi
 fi
 
 echo "------------------------------------------------"
@@ -144,18 +147,20 @@ if [ "${PLUGINS}" == "" ]; then
 fi
 
 if [ "${AGGREGATOR_DIR}" != "" ]; then
-    echo ""
-    echo "Assembling aggregator plugin dependencies..."
-    echo ""
-    pushd "${AGGREGATOR_DIR}"
-    mvn hpi:assemble-dependencies
-    if [ $? != 0 ];then
-        echo "*****"
-        echo "***** Error assembling dependencies from aggregator plugin. Maybe you need to rebuild everything."
-        echo "*****"
-        exit 1
+    if [ ! -f "${AGGREGATOR_DIR}/.pre-assembly" ]; then
+        echo ""
+        echo "Assembling aggregator plugin dependencies..."
+        echo ""
+        pushd "${AGGREGATOR_DIR}"
+        mvn hpi:assemble-dependencies
+        if [ $? != 0 ];then
+            echo "*****"
+            echo "***** Error assembling dependencies from aggregator plugin. Maybe you need to rebuild everything."
+            echo "*****"
+            exit 1
+        fi
+        popd
     fi
-    popd
 
     AGGREGATOR_ENV="PLUGINS_DIR=${AGGREGATOR_DIR}/target/plugins"
 fi
